Let the Elasticsearch exporter scrape /_cluster/health.
Right now this exporter only seems to report node specific data, but not
any cluster-level information. This data is returned under a separate
URI. Simply convert (almost) all of the data returned into Prometheus
metrics.

The metric descriptions are based on official Elasticsearch
documentation. It seems that some of these metrics aren't documented.
Proper strings for those should still be added.

The APIs of the client library used by this change differ from the ones
we use for the node-specific stats. MustNewConstMetric() is the
preferred way of creating constant values nowadays. We only need to keep
track of description objects, which we can allocate globally.